requirements.txt
Flask==2.3.3
flask-cors==3.1.1
torch==2.1.0
torchvision==0.15.1
Pillow==9.5.0
requests==2.31.0
